fix(coreaudio): use u32 instead of usize for size parameter
C function expects UInt32 (c_uint), not usize. On arm64, usize is
64-bit which doesn't match the 32-bit C parameter.

+++ zig-april-captions/src/coreaudio.zig
@@ -86,7 +86,7 @@ pub const AudioCapture = struct {
         if (!is_macos) return error.DeviceNotFound;
 
         var device_id: AudioDeviceID = 0;
-        var size: usize = @sizeOf(AudioDeviceID);
+        var size: u32 = @sizeOf(AudioDeviceID);
 
         const prop_addr = extern struct {
             mSelector: u32,
